+++ web-vue/src/pages/node/ssh.vue
@@ -127,6 +127,9 @@
           <a-form-model-item label="安装路径" prop="path">
             <a-input v-model="tempNode.path" placeholder="安装路径" />
           </a-form-model-item>
+           <a-form-model-item label="等待次数" prop="waitCount">
+            <a-input v-model="tempNode.waitCount" placeholder="上传插件端后,等待插件端启动成功次数，1次5秒。默认5次" />
+          </a-form-model-item>
           <a-form-model-item label="安装文件">
             <div class="clearfix">
               <a-upload :file-list="fileList" :remove="handleRemove" :before-upload="beforeUpload" accept=".zip">
@@ -498,6 +501,7 @@ export default {
         formData.append("id", this.temp.id);
         formData.append("nodeData", JSON.stringify({ ...this.tempNode }));
         formData.append("path", this.tempNode.path);
+        formData.append("waitCount",this.tempNode.waitCount);
         // 提交数据
         installAgentNode(formData).then((res) => {
           if (res.code === 200) {